Minor syconn backend changes to speedup the server
The connectivity graph initialization in SyConnBackend is
suppressed to improve the server overhead. The working directories are
set first to start the server and load the SyConnBackend and
KnossosDataset. Organelle command line argument is now integrated in
the Neuroglancer client. The int2str_converter is modified to handle
KeyError exceptions.