Update dynamic properties
The threshold mask for total path length is now optional. Recently
added `j0126` dataset does not currently have the compartment path
lengths cached. Hence, all the cells are displayed in the inline
properties.

The viewer state can now be reset any time using `ctrl+x`. This is
different from Neuroglancer's native `x` in that it additionally sets
the viewer position to the center of the dataset.